# go-backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.Company:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.HTTPError:
    properties:
      message:
        example: Error description
        type: string
    type: object
...
  services.CompanyTableResponse:
    properties:
      display_name:
        type: string
      id:
        type: integer
      name:
        type: string
      table_type:
        type: string
    type: object
  services.CompanyUnloadPlaceResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  services.CompanyUserRequest:
    properties:
      is_primary:
        type: boolean
      required_approval:
        type: boolean
      username:
        type: string
    type: object
  services.CompanyUserResponse:
    properties:
      first_name:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      last_name:
        type: string
      middle_name:
        type: string
      position:
        type: string
      required_approval:
        type: boolean
      username:
        type: string
    type: object
  services.CompanyWithUsersExtendedResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      unload_places:
        items:
          $ref: '#/definitions/services.CompanyUnloadPlaceResponse'
        type: array
      user_count:
        type: integer
    type: object
  services.CompanyWithUsersResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      user_count:
        type: integer
    type: object
...
  services.CreateCompanyRequest:
    properties:
      name:
        type: string
    type: object
...
  services.UpdateCompanyTablesRequest:
    properties:
      table_ids:
        items:
          type: integer
        type: array
    type: object
  services.UpdateCompanyUnloadPlacesRequest:
    properties:
      unload_place_ids:
        items:
          type: integer
        type: array
    type: object
  services.UpdateCompanyUsersRequest:
    properties:
      users:
        items:
          $ref: '#/definitions/services.CompanyUserRequest'
        type: array
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: API системы управления пропусками (Бюро пропусков)
  title: Systemburo API
  version: "1.0"
paths:
...
  /companies:
    get:
      consumes:
      - application/json
      description: Возвращает список всех компаний, отсортированных по имени
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Company'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить все компании
      tags:
      - companies
    post:
      consumes:
      - application/json
      description: Создаёт новую компанию. Требуются права buropropuskov
      parameters:
      - description: Данные новой компании
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Создать компанию
      tags:
      - companies
  /companies/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет компанию по ID. Нельзя удалить если есть привязанные пользователи.
        Требуются права buropropuskov
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Удалить компанию
      tags:
      - companies
    put:
      consumes:
      - application/json
      description: Обновляет название компании по ID. Требуются права buropropuskov
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные компании
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновить компанию
      tags:
      - companies
  /companies/{id}/tables:
    get:
      consumes:
      - application/json
      description: Возвращает список активных таблиц, привязанных к компании
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CompanyTableResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить таблицы компании
      tags:
      - companies
    put:
      consumes:
      - application/json
      description: Обновляет привязку таблиц к компании. Требуются права buropropuskov
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: integer
      - description: Список ID таблиц
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateCompanyTablesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновить таблицы компании
      tags:
      - companies
  /companies/{id}/unload-places:
    get:
      consumes:
      - application/json
      description: Возвращает список активных мест разгрузки, привязанных к компании
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CompanyUnloadPlaceResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить места разгрузки компании
      tags:
      - companies
    put:
      consumes:
      - application/json
      description: Обновляет привязку мест разгрузки к компании. Требуются права buropropuskov
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: integer
      - description: Список ID мест разгрузки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateCompanyUnloadPlacesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновить места разгрузки компании
      tags:
      - companies
  /companies/{id}/users:
    get:
      consumes:
      - application/json
      description: Возвращает список ответственных пользователей компании
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CompanyUserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить пользователей компании
      tags:
      - companies
    put:
      consumes:
      - application/json
      description: Заменяет список ответственных пользователей компании с поддержкой
        обязательного согласования
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: integer
      - description: Список ответственных пользователей
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateCompanyUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновить пользователей компании
      tags:
      - companies
  /companies/with-users:
    get:
      consumes:
      - application/json
      description: Возвращает список компаний с количеством привязанных пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CompanyWithUsersResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить компании с количеством пользователей
      tags:
      - companies
  /companies/with-users-extended:
    get:
      consumes:
      - application/json
      description: Возвращает компании с количеством пользователей и местами разгрузки
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CompanyWithUsersExtendedResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить компании с расширенной информацией
      tags:
      - companies
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"